+++ components/atoms/NavbarTitle.tsx
@@ -1,6 +1,6 @@
 import Link from 'next/link';
 import { metaDataForBlog } from '../../data/meta';
-import BouzuyaImg from '../../../images/Bouzuya.svg';
+import BouzuyaImg from '../../images/Bouzuya.svg';
 
 const { title } = metaDataForBlog;
